+++ digiwf-libs/digiwf-spring-cloudstream-utils/digiwf-spring-cloudstream-utils-core/src/main/java/io/muenchendigital/digiwf/spring/cloudstream/utils/api/streaming/incident/service/IncidentService.java
@@ -17,7 +17,7 @@ import static io.muenchendigital.digiwf.spring.cloudstream.utils.api.streaming.i
 @RequiredArgsConstructor
 public class IncidentService {
 
-    private static final String MESSAGE_TYPE = "incident";
+    private static final String MESSAGE_TYPE = "createIncident";
 
     private final Sinks.Many<Message<String>> incidentSink;
 
@@ -26,13 +26,14 @@ public class IncidentService {
      *
      * @param messageHeaders incoming message headers
      * @param errorMessage   the specific error message
-     * @return
+     * @return success flag
      */
     public boolean sendIncident(final MessageHeaders messageHeaders, final String errorMessage) {
         final Message<String> message = MessageBuilder
                 .withPayload(errorMessage)
                 .setHeader(StreamingHeaders.TYPE, MESSAGE_TYPE)
                 .setHeader(DIGIWF_PROCESS_INSTANCE_ID, messageHeaders.get(DIGIWF_PROCESS_INSTANCE_ID))
+                // bpmn token is waiting on that message receive event. Thats where we want our incident to be placed:
                 .setHeader(DIGIWF_MESSAGE_NAME, messageHeaders.get(DIGIWF_MESSAGE_NAME))
                 .build();